1) Configuration System Overhaul: ConfigLoader completely rewritten to use environment-based configuration files (settings-dev.yaml, settings-prod.yaml) 2) Database Layer Modernization: Changed from DatabaseUtilities to DatabaseFacade throughout the codebase 3) Package Structure Cleanup: Removed __all__ declarations from multiple __init__.py files 4) Daily Job Architecture: Major refactoring to use dependency injection with shared database connection 5) Test Updates: Extensive test updates to match new architecture. Overall coverage > 90%

+++ db/create_database.sql
@@ -41,7 +41,7 @@ create table stock_prices (
     high REAL,
     low REAL,
     close REAL,
-    volume BIGINT,
+    volume NUMERIC(20,0),
     adj_close REAL
 ) tablespace stockie_data_ts;
 
@@ -94,7 +94,7 @@ create table stock_prices (
     high REAL,
     low REAL,
     close REAL,
-    volume BIGINT,
+    volume NUMERIC(20,0),
     adj_close REAL
 ) tablespace stockie_dev_data_ts;
